Fix hiding Book / Journal when there is an error message
When there is an error message (e.g. when user clicks REQUEST when the title is not filled in) an extra div appears at the top of the form, which messes up the first-of-type selector.

We try to workaround the issue by switching to the last-of-type selector instead, which matches the CANCEL / REQUEST buttons at the bottom.